scripts/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
Jinja2==3.1.4
markdownify==0.13.1
MarkupSafe==3.0.2
nodeenv==1.9.1
prisma==0.15.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
StrEnum==0.4.15
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.2
